# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "List of all comments",
                "operationId": "6f6135f69be6fdcf9f2d5a662982c107",
                "responses": {
                    "200": {
                        "description": "List of albums",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "postId": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Lorem ipsum"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "example@email.com"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "Blanditiis corporis dolore dolorem eaque fuga"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create a new comment",
                "operationId": "893295f958c5d7af0be6146162c74990",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "postId": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "New comment"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "newexample@email.com"
                                            },
                                            "body": {
                                                "type": "string",
                                                "example": "New post"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "postId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New comment"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "newexample@email.com"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "New post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{comment}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Show a specific comment",
                "operationId": "bb1f108eef9515d0d7749f362aa04e76",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment's id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "postId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New comment"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "newexample@email.com"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "New post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{id}": {
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Update a specific comment",
                "operationId": "6867492787de7ddfdd8b49a22c2ef812",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the comment to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Data for updating the comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "postId": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Some name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example@email.com"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "postId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "example@email.com"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "status code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "Validation error(s)": {
                                            "type": "object",
                                            "additionalProperties": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comment not found or something went wrong",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}